+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.5.12] - 2021-03-09
+
+- Remove unnecessary Box from WebSocket messages.
+- Export subscription type to Federation SDL. (for [GraphGate](https://github.com/async-graphql/graphgate) 😁)
+- Add `extends` attribute for derive macros Subscription and MergedSubscription.
+- Add `SchemaBuilder::override_name` method. [#437](https://github.com/async-graphql/async-graphql/issues/437)
+
 ## [2.5.11] - 2021-03-07
 
 - Execute `_entity` requests in parallel. [#431](https://github.com/async-graphql/async-graphql/issues/431)